Workshop-3/python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
pytest==7.4.3
httpx==0.25.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
